# back-end/UniFood/src/main/webapp/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "UniFood",
    "description" : "API REST do UniFood",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/UniFood/rest",
    "variables" : { }
  } ],
...
  "paths" : {
...
    "/avaliacao/consulta" : {
      "post" : {
        "tags" : [ "Avaliações" ],
        "operationId" : "consulta",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AvaliacaoED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/avaliacao/consulta/{id}" : {
      "get" : {
        "tags" : [ "Avaliações" ],
        "operationId" : "consulta_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/avaliacao/removelista" : {
      "post" : {
        "tags" : [ "Avaliações" ],
        "operationId" : "removeLista",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/AvaliacaoED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/frequentacao/consulta" : {
      "post" : {
        "tags" : [ "Frequentações" ],
        "operationId" : "consulta_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FrequentacaoED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/frequentacao/consulta/{id}" : {
      "get" : {
        "tags" : [ "Frequentações" ],
        "operationId" : "consulta_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/frequentacao/removelista" : {
      "post" : {
        "tags" : [ "Frequentações" ],
        "operationId" : "removeLista_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/FrequentacaoED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/funcao/consulta" : {
      "post" : {
        "tags" : [ "Funcões" ],
        "operationId" : "consulta_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FuncaoED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/funcao/consulta/{id}" : {
      "get" : {
        "tags" : [ "Funcões" ],
        "operationId" : "consulta_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/funcao/removelista" : {
      "post" : {
        "tags" : [ "Funcões" ],
        "operationId" : "removeLista_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/FuncaoED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/nota/consulta" : {
      "post" : {
        "tags" : [ "Notas" ],
        "operationId" : "consulta_6",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NotaED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/nota/consulta/{id}" : {
      "get" : {
        "tags" : [ "Notas" ],
        "operationId" : "consulta_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/nota/removelista" : {
      "post" : {
        "tags" : [ "Notas" ],
        "operationId" : "removeLista_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/NotaED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/restaurante/consulta" : {
      "post" : {
        "tags" : [ "Restaurantes" ],
        "operationId" : "consulta_8",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RestauranteED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/restaurante/consulta/{id}" : {
      "get" : {
        "tags" : [ "Restaurantes" ],
        "operationId" : "consulta_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/restaurante/removelista" : {
      "post" : {
        "tags" : [ "Restaurantes" ],
        "operationId" : "removeLista_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/RestauranteED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/universidade/altera" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "altera_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/consulta" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "consulta_10",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/consulta/{id}" : {
      "get" : {
        "tags" : [ "Universidade" ],
        "operationId" : "consulta_11",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/conta" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "conta_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/inclui" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "inclui_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/lista" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "lista_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/remove" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "remove_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UniversidadeED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/universidade/removelista" : {
      "post" : {
        "tags" : [ "Universidade" ],
        "operationId" : "removeLista_5",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/UniversidadeED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/usuario/consulta" : {
      "post" : {
        "tags" : [ "Usuarios" ],
        "operationId" : "consulta_12",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UsuarioED"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
    "/usuario/consulta/{id}" : {
      "get" : {
        "tags" : [ "Usuarios" ],
        "operationId" : "consulta_13",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
    "/usuario/removelista" : {
      "post" : {
        "tags" : [ "Usuarios" ],
        "operationId" : "removeLista_6",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/UsuarioED"
                }
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "AvaliacaoED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "codNota" : {
            "type" : "integer",
            "format" : "int32"
          },
          "emailUsuario" : {
            "type" : "string"
          },
          "codRU" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "FrequentacaoED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "emailUsuario" : {
            "type" : "string"
          },
          "siglaUniversidade" : {
            "type" : "string"
          }
        }
      },
      "FuncaoED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tipo" : {
            "type" : "string"
          },
          "nomeCurso" : {
            "type" : "string"
          },
          "periodo" : {
            "type" : "string"
          }
        }
      },
      "NotaED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "emailUsuario" : {
            "type" : "string"
          },
          "periodoNota" : {
            "type" : "string"
          },
          "notaEstrelas" : {
            "type" : "integer",
            "format" : "int32"
          },
          "comentario" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "string"
          },
          "duracaoNota" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RestauranteED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "siglaRU" : {
            "type" : "string"
          },
          "campus" : {
            "type" : "string"
          },
          "siglaUniversidade" : {
            "type" : "string"
          },
          "mediaAvaliacao" : {
            "type" : "string"
          }
        }
      },
      "UniversidadeED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "nome" : {
            "type" : "string"
          },
          "estadoUF" : {
            "type" : "string"
          },
          "cidade" : {
            "type" : "string"
          },
          "sigla" : {
            "type" : "string"
          }
        }
      },
      "UsuarioED" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "senha" : {
            "type" : "string"
          },
          "nome" : {
            "type" : "string"
          },
          "sexo" : {
            "type" : "string"
          },
          "preferenciaAlimentar" : {
            "type" : "string"
          },
          "tipo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "email" : {
            "type" : "string"
          }
        }
      }
    }
  }
}